+++ src/compas_viewer/components/renderer/camera.py
@@ -2,7 +2,6 @@ from math import atan2
 from math import radians
 from math import tan
 from typing import TYPE_CHECKING
-from typing import Callable
 from typing import Optional
 
 from numpy import array
@@ -17,29 +16,27 @@ from compas.geometry import Rotation
 from compas.geometry import Transformation
 from compas.geometry import Translation
 from compas.geometry import Vector
+from compas_viewer.base import Base
 
 if TYPE_CHECKING:
     # https://peps.python.org/pep-0484/#runtime-or-type-checking
-    from .renderer import Renderer
+    pass
 
 
 class Position(Vector):
-    """
-    The position of the camera.
-
-    Parameters
-    ----------
-    vector : tuple[float, float, float]
-        The position of the camera.
-    on_update : Callable
-        A callback function that is called when the position changes.
-
-    """
-
-    def __init__(self, vector: tuple[float, float, float], on_update: Optional[Callable] = None):
-        self.on_update = on_update
-        self.pause_update = True
+    def __init__(self, vector, on_update=None):
         super().__init__(*vector)
+        self.pause_update = False
+        if on_update:
+            self.on_update = on_update
+
+    def set(self, x, y, z, pause_update=False):
+        pause_update = pause_update or self.pause_update
+        if hasattr(self, "on_update") and not pause_update:
+            self.on_update([x, y, z])
+        self._x = x
+        self._y = y
+        self._z = z
 
     @property
     def x(self):
@@ -47,7 +44,7 @@ class Position(Vector):
 
     @x.setter
     def x(self, x):
-        if self.on_update is not None and not self.pause_update:
+        if hasattr(self, "on_update") and not self.pause_update:
             self.on_update([x, self.y, self.z])
         self._x = float(x)
 
@@ -57,7 +54,7 @@ class Position(Vector):
 
     @y.setter
     def y(self, y):
-        if self.on_update is not None and not self.pause_update:
+        if hasattr(self, "on_update") and not self.pause_update:
             self.on_update([self.x, y, self.z])
         self._y = float(y)
 
@@ -67,25 +64,16 @@ class Position(Vector):
 
     @z.setter
     def z(self, z):
-        if self.on_update is not None and not self.pause_update:
+        if hasattr(self, "on_update") and not self.pause_update:
             self.on_update([self.x, self.y, z])
         self._z = float(z)
-
-    def set(self, x: float, y: float, z: float, pause_update: bool = False):
-        """Set the position of the camera."""
-        pause_update = pause_update or self.pause_update
-        if self.on_update is not None and not pause_update:
-            self.on_update([x, y, z])
-        self._x = x
-        self._y = y
-        self._z = z
 
 
 class RotationEuler(Position):
     pass
 
 
-class Camera:
+class Camera(Base):
     """Camera object for the default view.
 
     Parameters
@@ -127,20 +115,33 @@ class Camera:
         The scale factor for camera's near, far and pan_delta.
     """
 
-    def __init__(self, renderer: "Renderer"):
-        self.renderer = renderer
-        self.config = renderer.config.camera
-        self._position = Position((0.0, 0.0, 10.0 * self.config.scale), on_update=self._on_position_update)
-        self._rotation = RotationEuler((0, 0, 0), on_update=self._on_rotation_update)
-        self._target = Position((0, 0, 0), on_update=self._on_target_update)
-        self._position.pause_update = False
-        self._rotation.pause_update = False
-        self._target.pause_update = False
-        # Camera position only modifiable in perspective view mode.
-        self.reset_position()
-        if self.renderer.config.viewmode == "perspective":
-            self.position = Position(self.config.position)
-        self.target = Position(self.config.target)
+    def __init__(
+        self,
+        fov: Optional[float] = 45.0,
+        near: Optional[float] = 0.1,
+        far: Optional[float] = 1000.0,
+        init_position: Optional[tuple] = [10.0, 10.0, 10.0],
+        init_target: Optional[tuple] = [0.0, 0.0, 0.0],
+        scale: Optional[float] = 1.0,
+        zoomdelta: Optional[float] = 0.05,
+        rotationdelta: Optional[float] = 0.01,
+        pan_delta: Optional[float] = 0.05,
+    ) -> None:
+        self.fov = fov
+        self.near = near
+        self.far = far
+        self.scale = scale
+        self.zoomdelta = zoomdelta
+        self.rotationdelta = rotationdelta
+        self.pan_delta = pan_delta
+
+        self._position = Position([0, 0, 10 * scale], on_update=self._on_position_update)
+        self._rotation = RotationEuler([0, 0, 0], on_update=self._on_rotation_update)
+        self._target = Position([0, 0, 0], on_update=self._on_target_update)
+
+        self.reset_position(viewmode="perspective")
+        self.target.set(*init_target)
+        self.position.set(*init_position)
 
     @property
     def position(self) -> Position:
@@ -303,16 +304,18 @@ class Camera:
         self.target.set(*target, pause_update=True)
         self.position.set(*position, pause_update=True)
 
-    def reset_position(self):
+    def reset_position(self, viewmode: Optional[str] = None):
         """Reset the position of the camera based current view type."""
         self.target.set(0, 0, 0, False)
-        if self.renderer.viewmode == "perspective":
+        viewmode = viewmode or self.viewer.renderer.viewmode
+
+        if viewmode == "perspective":
             self.rotation.set(pi / 4, 0, -pi / 4, False)
-        if self.renderer.viewmode == "top":
+        if viewmode == "top":
             self.rotation.set(0, 0, 0, False)
-        if self.renderer.viewmode == "front":
+        if viewmode == "front":
             self.rotation.set(pi / 2, 0, 0, False)
-        if self.renderer.viewmode == "right":
+        if viewmode == "right":
             self.rotation.set(pi / 2, 0, pi / 2, False)
 
     def rotate(self, dx: float, dy: float):
@@ -333,8 +336,8 @@ class Camera:
         is a perspective view (``camera.renderer.config.viewmode == "perspective"``).
 
         """
-        if self.renderer.config.viewmode == "perspective":
-            self.rotation += [-self.config.rotationdelta * dy, 0, -self.config.rotationdelta * dx]
+        if self.viewer.renderer.config.viewmode == "perspective":
+            self.rotation += [-self.rotationdelta * dy, 0, -self.rotationdelta * dx]
 
     def pan(self, dx: float, dy: float):
         """Pan the camera based on current mouse movement.
@@ -349,7 +352,7 @@ class Camera:
             with each increment the size of :attr:`Camera.pan_delta`.
         """
         R = Rotation.from_euler_angles(self.rotation)
-        T = Translation.from_vector([-dx * self.config.pan_delta * self.config.scale, dy * self.config.pan_delta * self.config.scale, 0])
+        T = Translation.from_vector([-dx * self.pan_delta * self.scale, dy * self.pan_delta * self.scale, 0])
         M = (R * T).matrix
         vector = [M[i][3] for i in range(3)]
         self.target += vector
@@ -364,7 +367,7 @@ class Camera:
             of :attr:`compas_viewer.components.renderer.Camera.config.zoomdelta`.
 
         """
-        self.distance -= steps * self.config.zoomdelta * self.distance
+        self.distance -= steps * self.zoomdelta * self.distance
 
     def projection(self, width: int, height: int) -> list[list[float]]:
         """Compute the projection matrix corresponding to the current camera settings.
@@ -387,14 +390,14 @@ class Camera:
 
         """
         aspect = width / height
-        if self.renderer.viewmode == "perspective":
-            P = self.perspective(self.config.fov, aspect, self.config.near * self.config.scale, self.config.far * self.config.scale)
+        if self.viewer.renderer.viewmode == "perspective":
+            P = self.perspective(self.fov, aspect, self.near * self.scale, self.far * self.scale)
         else:
             left = -self.distance
             right = self.distance
             bottom = -self.distance / aspect
             top = self.distance / aspect
-            P = self.ortho(left, right, bottom, top, self.config.near * self.config.scale, self.config.far * self.config.scale)
+            P = self.ortho(left, right, bottom, top, self.near * self.scale, self.far * self.scale)
         return list(asfortranarray(P, dtype=float32))
 
     def viewworld(self) -> list[list[float]]: