+++ src/compas_viewer/scene/__init__.py
@@ -2,13 +2,13 @@
 This package provides scene object plugins for visualizing COMPAS objects in `compas_viewer`.
 """
 
-from typing import Union
 from compas.scene import register
 from compas.plugins import plugin
 from compas.datastructures import Mesh
 from compas.datastructures import Graph
 from compas.geometry import (
     Point,
+    Pointcloud,
     Line,
     Vector,
     Circle,
@@ -25,13 +25,13 @@ from compas.geometry import (
     Polyhedron,
     Frame,
     NurbsSurface,
-    Geometry,
 )
 
 from .sceneobject import ViewerSceneObject
 from .meshobject import MeshObject
 from .graphobject import GraphObject
 from .pointobject import PointObject
+from .pointcloudobject import PointcloudObject
 from .lineobject import LineObject
 from .vectorobject import VectorObject
 from .tagobject import TagObject, Tag
@@ -47,9 +47,12 @@ from .cylinderobject import CylinderObject
 from .ellipseobject import EllipseObject
 from .coneobject import ConeObject
 from .capsuleobject import CapsuleObject
+from .geometryobject import GeometryObject
+from .groupobject import Group
+from .groupobject import GroupObject
+from .collectionobject import Collection
 from .polyhedronobject import PolyhedronObject
 from .collectionobject import CollectionObject
-from .geometryobject import GeometryObject
 
 
 @plugin(category="drawing-utils", requires=["compas_viewer"])
@@ -68,6 +71,7 @@ def register_scene_objects():
     register(Mesh, MeshObject, context="Viewer")
     register(Graph, GraphObject, context="Viewer")
     register(Point, PointObject, context="Viewer")
+    register(Pointcloud, PointcloudObject, context="Viewer")
     register(Line, LineObject, context="Viewer")
     register(Tag, TagObject, context="Viewer")
     register(Frame, FrameObject, context="Viewer")
@@ -83,8 +87,9 @@ def register_scene_objects():
     register(Ellipse, EllipseObject, context="Viewer")
     register(Cone, ConeObject, context="Viewer")
     register(Capsule, CapsuleObject, context="Viewer")
+    register(list, GroupObject, context="Viewer")
+    register(Collection, CollectionObject, context="Viewer")
     register(Polyhedron, PolyhedronObject, context="Viewer")
-    register(list[Union[Geometry, Mesh]], CollectionObject, context="Viewer")
 
     try:
         from compas_occ.brep import OCCBrep
@@ -137,4 +142,8 @@ __all__ = [
     "NurbsSurface",
     "NurbsSurfaceObject",
     "GeometryObject",
+    "Group",
+    "GroupObject",
+    "Collection",
+    "CollectionObject",
 ]