+++ src/compas_viewer/scene/sceneobject.py
@@ -1,4 +1,3 @@
-from typing import TYPE_CHECKING
 from typing import Any
 from typing import Optional
 
@@ -12,21 +11,18 @@ from compas.geometry import Transformation
 from compas.geometry import transform_points_numpy
 from compas.itertools import flatten
 from compas.scene import SceneObject
+from compas_viewer.base import Base
 from compas_viewer.components.renderer.shaders import Shader
 from compas_viewer.gl import make_index_buffer
 from compas_viewer.gl import make_vertex_buffer
 from compas_viewer.gl import update_index_buffer
 from compas_viewer.gl import update_vertex_buffer
 
-if TYPE_CHECKING:
-    from compas_viewer import Viewer
-
-
 # Type template of point/line/face data for generating the buffers.
 ShaderDataType = tuple[list[Point], list[Color], list[list[int]]]
 
 
-class ViewerSceneObject(SceneObject):
+class ViewerSceneObject(SceneObject, Base):
     """
     Base class for all Viewer scene objects
     which also includes the  GL buffer creation and drawing methods.
@@ -47,10 +43,10 @@ class ViewerSceneObject(SceneObject):
         Whether to show lines/edges of the object. Default is the value of `show_lines` in `viewer.config`.
     show_faces : bool, optional
         Whether to show faces of the object. Default is the value of `show_faces` in `viewer.config`.
-    lineswidth : float, optional
-        The line width to be drawn on screen. Default is the value of `lineswidth` in `viewer.config`.
-    pointssize : float, optional
-        The point size to be drawn on screen. Default is the value of `pointssize` in `viewer.config`.
+    linewidth : float, optional
+        The line width to be drawn on screen. Default is the value of `linewidth` in `viewer.config`.
+    pointsize : float, optional
+        The point size to be drawn on screen. Default is the value of `pointsize` in `viewer.config`.
     opacity : float, optional
         The opacity of the object. Default is the value of `opacity` in `viewer.config`.
     **kwargs : dict, optional
@@ -71,9 +67,9 @@ class ViewerSceneObject(SceneObject):
         Whether to show lines/edges of the object.
     show_faces : bool
         Whether to show faces of the object.
-    lineswidth : float
+    linewidth : float
         The line width to be drawn on screen
-    pointssize : float
+    pointsize : float
         The point size to be drawn on screen.
     opacity : float
         The opacity of the object.
@@ -91,38 +87,35 @@ class ViewerSceneObject(SceneObject):
 
     def __init__(
         self,
-        viewer: "Viewer",
         is_selected: bool = False,
         is_locked: bool = False,
         is_visible: bool = True,
         show_points: Optional[bool] = None,
         show_lines: Optional[bool] = None,
         show_faces: Optional[bool] = None,
-        lineswidth: Optional[float] = None,
-        pointssize: Optional[float] = None,
+        linewidth: Optional[float] = None,
+        pointsize: Optional[float] = None,
         opacity: Optional[float] = None,
         use_rgba: bool = False,
         **kwargs,
     ):
         #  Basic
         super().__init__(**kwargs)
-        self.viewer = viewer
-        self.scene = viewer.scene
-        self.renderer = viewer.renderer
         self.is_visible = is_visible
-        self.show_points = self.viewer.config.show_points if show_points is None else show_points
-        self.show_lines = self.viewer.config.show_lines if show_lines is None else show_lines
-        self.show_faces = self.viewer.config.show_faces if show_faces is None else show_faces
-        self.lineswidth = lineswidth or self.viewer.config.lineswidth
-        self.pointssize = pointssize or self.viewer.config.pointssize
-        self.opacity = opacity or self.viewer.config.opacity
+        self.show_points = show_points if show_points is not None else False
+        self.show_lines = show_lines if show_lines is not None else True
+        self.show_faces = show_faces if show_faces is not None else True
+        self.linewidth = linewidth if linewidth is not None else self.viewer.config.ui.display.linewidth
+        self.pointsize = pointsize if pointsize is not None else self.viewer.config.ui.display.pointsize
+        self.opacity = opacity if opacity is not None else self.viewer.config.ui.display.opacity
 
         #  Selection
         self._is_locked = is_locked
         self.is_selected = not is_locked and is_selected
-        self.instance_color = Color.from_rgb255(*next(self.scene._instance_colors_generator))
+        # TODO scene
+        self.instance_color = Color.from_rgb255(*next(self.viewer.scene._instance_colors_generator))
         if not is_locked:
-            self.scene.instance_colors[self.instance_color.rgb255] = self
+            self.viewer.scene.instance_colors[self.instance_color.rgb255] = self
 
         #  Visual
         self.background: bool = False
@@ -133,7 +126,6 @@ class ViewerSceneObject(SceneObject):
         self._matrix_buffer: Optional[list[list[float]]] = None
         self._bounding_box: Optional[list[float]] = None
         self._bounding_box_center: Optional[Point] = None
-        self._is_collection = False
 
         #  Primitive
         self._points_data: Optional[ShaderDataType] = None
@@ -154,6 +146,7 @@ class ViewerSceneObject(SceneObject):
         self._is_locked = value
         if value:
             self.is_selected = False
+            # scene parent
             self.scene.instance_colors.pop(self.instance_color.rgb255)
         else:
             self.scene.instance_colors[self.instance_color.rgb255] = self
@@ -347,7 +340,7 @@ class ViewerSceneObject(SceneObject):
             )
 
         #  Update the canvas.
-        self.renderer.update()
+        self.viewer.renderer.update()
 
     def _update_bounding_box(self, positions: Optional[list[Point]] = None):
         """Update the bounding box of the object"""
@@ -398,7 +391,7 @@ class ViewerSceneObject(SceneObject):
             shader.bind_attribute("position", self._lines_buffer["positions"])
             shader.bind_attribute("color", self._lines_buffer["colors"], step=4)
             shader.draw_lines(
-                width=self.lineswidth,
+                width=self.linewidth,
                 elements=self._lines_buffer["elements"],
                 n=self._lines_buffer["n"],
                 background=self.background,
@@ -409,7 +402,7 @@ class ViewerSceneObject(SceneObject):
             shader.bind_attribute("position", self._points_buffer["positions"])
             shader.bind_attribute("color", self._points_buffer["colors"], step=4)
             shader.draw_points(
-                size=self.pointssize,
+                size=self.pointsize,
                 elements=self._points_buffer["elements"],
                 n=self._points_buffer["n"],
                 background=self.background,
@@ -432,12 +425,12 @@ class ViewerSceneObject(SceneObject):
         # Points
         if self._points_buffer is not None and self.show_points:
             shader.bind_attribute("position", self._points_buffer["positions"])
-            shader.draw_points(size=self.pointssize, elements=self._points_buffer["elements"], n=self._points_buffer["n"])
+            shader.draw_points(size=self.pointsize, elements=self._points_buffer["elements"], n=self._points_buffer["n"])
         # Lines
         if self._lines_buffer is not None and (self.show_lines or wireframe):
             shader.bind_attribute("position", self._lines_buffer["positions"])
             shader.draw_lines(
-                width=self.lineswidth + self.renderer.selector.PIXEL_SELECTION_INCREMENTAL,
+                width=self.linewidth + self.viewer.renderer.selector.PIXEL_SELECTION_INCREMENTAL,
                 elements=self._lines_buffer["elements"],
                 n=self._lines_buffer["n"],
             )